CalcOS

+++ arch/x86/cpu/idt/sys_write.c
@@ -9,6 +9,6 @@
 #include <sound.h>
 #include <forth.h>
 
-int sys_write(int fd, const char* str, uint8_t color) {
+int sys_write(int fd, const char* str, uint32_t color) {
     return (int)_syscall(SYS_WRITE, (uint32_t)fd, (uint32_t)str, (uint32_t)color);
 }
