+++ include/utils.h
@@ -10,7 +10,7 @@ int atoi(char *str);
 void itoa(int n, char* s);
 void htoa(int n, char str[]);
 
-void draw_button(int _x, int _y, int _width, int _height, const char *_msg, uint8_t color, uint8_t text_color);
+void draw_button(int _x, int _y, int _width, int _height, const char *_msg, uint32_t color, uint32_t text_color);
 
 uint8_t check_battery();
 uint8_t bcd_to_bin(uint8_t val);
